PWX-30230 : Create all px configmaps in same namespace as portworx
