+++ k8s/core/configmap/types.go
@@ -58,7 +58,14 @@ var (
 type FatalCb func(format string, args ...interface{})
 
 type configMap struct {
+	config   *coreConfigMap
+	pxNs     string
+	copylock *coreConfigMap
+}
+
+type coreConfigMap struct {
 	name                   string
+	nameSpace              string
 	kLockV1                k8sLock
 	kLocksV2Mutex          sync.Mutex
 	kLocksV2               map[string]*k8sLock
